backend/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
python-dotenv
python-multipart
openpyxl
xlrd
